finish FanOutHandler-based implementation
Really it's a bit of a coincidence that the FanOutHandler can assume that any
error thrown by the fetcher means the cluster was unreachable -- after all,
not so long ago the fetcher would have thrown errors on non-200 statuses! In
those days, this kind of logic would be ignoring valuable information from an
error.

So who's to say there aren't other valuable error cases that we're destroying
here?

So I will probably rewrite this whole thing, but here's an approach that does
technically work.

Signed-off-by: Jamie Klassen <jklassen@vmware.com>