special treatment for cluster-level error case
this is better than an obsessively-specific "cluster unreachable" error that
is all mixed up with the other errors that have a lot to do with resource
fetching, but it might be a bit too precious with the type system -- this
changes the wire format, the notion of a "cluster with errors" in a few places.

partly I like this because we're effectively using exceptions as control flow
here and this emphasizes the special treatment of this case, but it also feels
like it changes more parts of the codebase than strictly necessary.

Along the way I tried having the .errors field on the ClusterObjects type be
an array of ClusterError | ResourceFetchError and that seemed to result in a
smaller diff overall, so I may try returning to that.

Signed-off-by: Jamie Klassen <jklassen@vmware.com>