backfill test for loading default kubeconfig
Strictly speaking, the `loadFromDefault` method on `KubeConfig` as it is
being called in `KubernetesClientProvider` has many other behaviours that this
test does not exercise -- it has logic for reading kubeconfig files (based on
an env var or default location in a home directory), provisions for running on
Windows, and even contacting an apiserver on localhost.

I'm making a breaking assumption here that these other scenarios are not
important for Backstage users, and the one that really matters is the
in-cluster one described in this test.

Signed-off-by: Jamie Klassen <jklassen@vmware.com>