SPIKE: Microsoft manages multiple tokens
One per audience -- the "refresh" experience here has a lot of rough edges to
consider, and the choice to treat whatever scope is requested as the "default"
is probably a silly one. Nonetheless, this code shows some of the relevant
places to write tests while driving out a proper implementation.

Signed-off-by: Jamie Klassen <jklassen@vmware.com>